content/canvas/test/test_fallback.basic.html

<!DOCTYPE HTML>
<title>Canvas test: fallback.basic</title>
<script src="/MochiKit/MochiKit.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>
var _deferred = false;

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

ok(canvas.childNodes.length == 1, "canvas.childNodes.length == 1");

if (!_deferred) SimpleTest.finish();
});
</script>
